java_modernization/requirements.txt
openai
loguru
python-dotenv
streamlit
google-generativeai
streamlit-option-menu
ollama
tree-sitter
tree-sitter-java
rich
neo4j
pypandoc
python-docx 
markdown
beautifulsoup4